update test snapshots
