+++ rpc/mod.ts
@@ -3,5 +3,4 @@ export * from "./Error.ts";
 export * from "./messages.ts";
 export * from "./smoldot.ts";
 export * from "./types/mod.ts";
-export * from "./util.ts";
 export * from "./ws.ts";
